Move QLed to widgets directory
