+++ Bricks/widgets/QLed/QLed.py
