+++ Bricks/widgets/QLed/qled.qrc
