+++ Bricks/widgets/QLed/qled_resources.py
