+++ Bricks/widgets/QLed/resources/qled.png
