+++ Bricks/widgets/QLed/resources/screenshot.png
