+++ Bricks/widgets/QLed/resources/screenshot2.png
