+++ Bricks/widgets/QLed/resources_readme.rst
